warning fix

+++ src/gui/Qt5/ui/mainwindow.cpp
@@ -125,13 +125,13 @@ void MainWindow::createMenus()
 {
     //reattach items to "Windows" menu
 
-    for(int i = 0; i < m_views.size(); i++)
+    for(size_t i = 0; i < m_views.size(); i++)
     {
         IView* view = m_views[i];
         const QString title = view->getName();
         QAction* action = ui->menuWindows->addAction(title);
 
-        action->setData(i);
+        action->setData(static_cast<int>(i));
         connect(ui->menuWindows, SIGNAL(triggered(QAction *)), this, SLOT(activateWindow(QAction*)));
     }
 }
